market_data_service/requirements.txt
pandas
psycopg2-binary
python-binance
asyncio
websockets
python-dotenv
